requirements.txt
numpy
tensorflow
pillow
scipy
keras
idx2numpy